Creación productos. Cambios esteticos en admin y cambio en vista productos con descuentos

+++ src/database/models/Sizes.js
@@ -8,7 +8,7 @@ module.exports = (sequelize, DataTypes) =>{
             allowNull: false,
             autoIncrement: true
         },
-        number: DataTypes.INTEGER
+        number: DataTypes.DECIMAL
     };
